Restore correct value of combobox after the combobox has been built
Before was done in onCfg_renderModeChanged but often the combobox was still empty and it was not possible to restore the value.

If with the current solution fails because the cfg_renderMode is not available when the combobox is built, the function to restore the value must be called both from the onCompleted and from onCfg_renderModeChanged.